+++ app/server-actions/movies/add-movie.ts
@@ -5,6 +5,8 @@ import { createServerActionClient } from "@supabase/auth-helpers-nextjs";
 import { cookies } from "next/headers";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
+import { createClient } from "@/utils/supabase/server";
+import { prisma } from "@/lib/prisma";
 
 const MAX_FILE_SIZE = 5000000;
 
@@ -16,6 +18,17 @@ const ACCEPTED_IMAGE_TYPES = [
 ];
 
 export async function addMovie(prevState: any, formData: FormData) {
+  const supabase = createClient();
+  const { data, error } = await supabase.auth.getUser();
+
+  if (error || !data?.user) {
+    return {
+      type: "error",
+      message: "You must be connected to add a movie",
+      errors: null,
+    };
+  }
+
   const schema = z.object({
     title: z
       .string()
@@ -27,7 +40,7 @@ export async function addMovie(prevState: any, formData: FormData) {
     description: z
       .string()
       .min(5, "Le synopsis doit faire au moins 5 caractères"),
-    release_date: z.number().min(4, "L'année de sortie est obligatoire"),
+    release_date: z.string().min(4, "L'année de sortie est obligatoire"),
     runtime: z.number().min(4, "L'année de sortie est obligatoire"),
     country_id: z.string().min(1, "Le pays est obligatoire"),
     genre_id: z.string().min(1, "Le genre est obligatoire"),
@@ -48,7 +61,7 @@ export async function addMovie(prevState: any, formData: FormData) {
     title: formData.get("title"),
     director_name: formData.get("director_name"),
     description: formData.get("description"),
-    release_date: Number(formData.get("release_date")),
+    release_date: formData.get("release_date"),
     runtime: Number(formData.get("runtime")),
     country_id: formData.get("country_id"),
     genre_id: formData.get("genre_id"),
@@ -98,111 +111,63 @@ export async function addMovie(prevState: any, formData: FormData) {
       };
     }
 
-    // movie insert
-    const { data: movieData, error: movieError } = await supabase
-      .from("movies")
-      .insert({
-        title,
-        release_date,
-        runtime,
-        description,
-        image_url: imageData?.path,
-      })
-      .select("id"); // get id movie
-
-    if (movieError) {
-      return {
-        type: "error",
-        message: "Erreur avec la base de données : Echec de l'ajout du film",
-      };
-    }
-
-    const movieId = movieData[0]?.id; // get movie id
-
-    // insert country in movie_countries
-    const countryIds = country_id.split(",").map(Number); // country_id on list format
-    const countryInsert = countryIds.map((id) => ({
-      movie_id: movieId,
-      country_id: id,
-    }));
-
-    const { error: countryError } = await supabase
-      .from("movie_countries")
-      .insert(countryInsert);
-
-    if (countryError) {
-      return {
-        type: "error",
-        message: "Erreur lors de l'insertion des pays",
-      };
-    }
-
-    //insert director
-    const { data: directorData, error: directorError } = await supabase
-      .from("directors")
-      .upsert({
-        name: director_name,
-      })
-      .select("id");
-
-    if (directorError) {
-      return {
-        type: "error",
-        message: "Erreur lors de l'insertion du réalisateur",
-      };
-    }
-
-    const directorId = directorData[0]?.id;
-    // Link director to movie
-    const { error: movieDirectorError } = await supabase
-      .from("movie_directors")
-      .insert({
-        movie_id: movieId,
-        director_id: directorId,
+    const result = await prisma.$transaction(async (prisma) => {
+      // Upsert director
+      const director = await prisma.directors.upsert({
+        where: { id: 0 }, // Placeholder
+        // where: { name: director_name }, // Only works if 'name' is defined as @unique in schema
+        update: { name: director_name },
+        create: { name: director_name },
       });
 
-    if (movieDirectorError) {
-      return {
-        type: "error",
-        message: "Erreur lors de la liaison du réalisateur au film",
-      };
-    }
+      // Create movie
+      const movie = await prisma.movies.create({
+        data: {
+          title,
+          release_date,
+          runtime,
+          description,
+          image_url: imageData?.path,
+        },
+      });
 
-    // insert genres in movie_genres
-    const genreIds = genre_id.split(",").map(Number); // genre_id on list format
-    const genreInserts = genreIds.map((id) => ({
-      movie_id: movieId,
-      genre_id: id,
-    }));
+      // Insert movie countries
+      const countryIds = country_id.split(",").map(Number);
+      await prisma.movie_countries.createMany({
+        data: countryIds.map((countryId) => ({
+          movie_id: movie.id,
+          country_id: countryId,
+        })),
+      });
 
-    const { error: genreError } = await supabase
-      .from("movie_genres")
-      .insert(genreInserts);
+      // Link director to movie
+      await prisma.movie_directors.create({
+        data: {
+          movie_id: movie.id,
+          director_id: director.id,
+        },
+      });
 
-    if (genreError) {
-      return {
-        type: "error",
-        message: "Erreur lors de l'insertion des genres",
-      };
-    }
+      // Insert genres
+      const genreIds = genre_id.split(",").map(Number);
+      await prisma.movie_genres.createMany({
+        data: genreIds.map((genreId) => ({
+          movie_id: movie.id,
+          genre_id: genreId,
+        })),
+      });
 
-    // insert keywords into movie_keywords
-    const keywordIds = keyword_id.split(",").map(Number);
-    const keywordInserts = keywordIds.map((id) => ({
-      movie_id: movieId,
-      keyword_id: id,
-    }));
+      // Insert keywords
+      const keywordIds = keyword_id.split(",").map(Number);
+      await prisma.movie_keywords.createMany({
+        data: keywordIds.map((keywordId) => ({
+          movie_id: movie.id,
+          keyword_id: keywordId,
+        })),
+      });
 
-    const { error: keywordError } = await supabase
-      .from("movie_keywords")
-      .insert(keywordInserts);
-
-    if (keywordError) {
-      return {
-        type: "error",
-        message: "Erreur lors de l'insertion des mots-clés",
-      };
-    }
+      return movie;
+    });
   } catch (error) {
     console.error("Error", error);
     return {
@@ -210,7 +175,6 @@ export async function addMovie(prevState: any, formData: FormData) {
       message: "Erreur avec la base de données : Echec de l'ajout du film",
     };
   }
-
   // redirect
 
   revalidatePath("/");
